+++ packages/aws-amplify-react/src/Auth/AuthStateWrapper.jsx
@@ -1,4 +1,5 @@
-import React, { Component } from 'react';
+import * as React from 'react';
+import { Component } from 'react';
 import Amplify, { ConsoleLogger as Logger } from '@aws-amplify/core';
 import Auth from '@aws-amplify/auth';
 
@@ -67,7 +68,7 @@ export default class AuthStateWrapper extends Component {
                     onStateChange: this.handleStateChange,
                     onAuthEvent: this.handleAuthEvent
                 });
-            })
+            });
 
         return (
             <div className="amplify-state-wrapper" style={theme.stateWrapper}>