+++ packages/aws-amplify-react/CHANGELOG.md
@@ -3,6 +3,119 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.2-unstable.2"></a>
+## [2.1.2-unstable.2](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.1.2-unstable.1...aws-amplify-react@2.1.2-unstable.2) (2018-10-25)
+
+
+### Features
+
+* **aws-amplify-react:** federated sign out methods refactor ([9edbd6e](https://github.com/aws/aws-amplify/commit/9edbd6e))
+
+
+
+
+<a name="2.1.2-unstable.1"></a>
+## [2.1.2-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.1.2-unstable.0...aws-amplify-react@2.1.2-unstable.1) (2018-10-25)
+
+
+### Bug Fixes
+
+* **aws-amplify-react aws-amplify-react-native:** Connect component ([#1868](https://github.com/aws/aws-amplify/issues/1868)) ([8dd6b55](https://github.com/aws/aws-amplify/commit/8dd6b55))
+
+
+
+
+<a name="2.1.2-unstable.0"></a>
+## [2.1.2-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.1.1...aws-amplify-react@2.1.2-unstable.0) (2018-10-23)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="2.1.1"></a>
+## [2.1.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.0.8-unstable.2...aws-amplify-react@2.1.1) (2018-10-17)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="2.0.8-unstable.2"></a>
+## [2.0.8-unstable.2](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.0.8-unstable.1...aws-amplify-react@2.0.8-unstable.2) (2018-10-16)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="2.0.8-unstable.1"></a>
+## [2.0.8-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.0.8-unstable.0...aws-amplify-react@2.0.8-unstable.1) (2018-10-08)
+
+
+### Bug Fixes
+
+* **aws-amplify-react:** jump to the initial state if not signed in ([d8779eb](https://github.com/aws/aws-amplify/commit/d8779eb))
+
+
+
+
+<a name="2.0.8-unstable.0"></a>
+## [2.0.8-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.0.7-unstable.1...aws-amplify-react@2.0.8-unstable.0) (2018-10-05)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="2.0.7"></a>
+## [2.0.7](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.0.7-unstable.1...aws-amplify-react@2.0.7) (2018-10-04)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="2.0.7-unstable.1"></a>
+## [2.0.7-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.0.7-unstable.0...aws-amplify-react@2.0.7-unstable.1) (2018-10-03)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="2.0.7-unstable.0"></a>
+## [2.0.7-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.0.6-unstable.1...aws-amplify-react@2.0.7-unstable.0) (2018-10-03)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="2.0.6"></a>
+## [2.0.6](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.0.6-unstable.1...aws-amplify-react@2.0.6) (2018-10-03)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="2.0.6-unstable.1"></a>
+## [2.0.6-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.0.6-unstable.0...aws-amplify-react@2.0.6-unstable.1) (2018-10-02)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
+<a name="2.0.6-unstable.0"></a>
+## [2.0.6-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.0.5...aws-amplify-react@2.0.6-unstable.0) (2018-10-02)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-react
+
 <a name="2.0.5"></a>
 ## [2.0.5](https://github.com/aws/aws-amplify/compare/aws-amplify-react@2.0.5-unstable.1...aws-amplify-react@2.0.5) (2018-09-27)